LabelMapper: Add more template instantiations for different dtype combinations.
This was originally introduced via b2d1fde33efa062caf41c790fec19062c774bb1d, but reverted in 5264b6c49540fb2a1b411d5490122a761aad140a in an attempt to simplify the code while hunting for a bug.
The bug turned out to be unrelated, but this change was not reintroduced (until now).